Add some locks on any possible cross thread access to the known tests, removes coroutine again.
This continues to be a cat and mouse game trying to find the cause of the hang. In both coroutine and task scenarios something goes wrong todo with loading/unloading a scene. Given we know of serialization crashes we are suspect of that. Going to remove the formerly serilaized attribute for now.